requirements_explicit.txt
chex==0.1.86
diffrax==0.6.0
equinox==0.11.4
h5py==3.11.0
jax==0.4.31
jax-cuda12-pjrt==0.4.31
jax-cuda12-plugin==0.4.31
jaxlib==0.4.31
jaxtyping==0.2.33
neptune==1.11.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.0.22
nvidia-cuda-cupti-cu12==12.6.37
nvidia-cuda-nvcc-cu12==12.6.20
nvidia-cuda-runtime-cu12==12.6.37
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.6.28
nvidia-cusolver-cu12==11.6.4.38
nvidia-cusparse-cu12==12.5.2.23
nvidia-nccl-cu12==2.22.3
nvidia-nvjitlink-cu12==12.6.20
opt-einsum==3.3.0
optax==0.2.3
pandas==2.2.2
PyYAML==6.0.2
ray==2.34.0
